requirements.txt
Flask==1.1.1
Flask-WTF==0.14.2
pycryptodome==3.9.6
werkzeug==0.16.0